+++ src/components/CartbarItem/index.scss
@@ -1,0 +1,168 @@
+.cartbarItem {
+    .order {
+        padding: 10px;
+        background: #eee;
+
+        .product {
+            margin: 0;
+            display: flex;
+
+            .photo {
+                width: 120px;
+                height: 160px;
+
+                .image {
+                    width: 100%;
+                    height: 100%;
+                    object-fit: cover;
+                }
+            }
+
+            .information {
+                height: 160px;
+                width: 150px;
+
+                line-height: 20px;
+                color: #000;
+
+                .name {
+                    white-space: nowrap;
+                    overflow: hidden;
+                    text-overflow: ellipsis;
+                }
+
+                .description {
+                    font-size: 16px;
+                    font-weight: 400;
+
+                    overflow: hidden;
+                    display: -webkit-box;
+                    -webkit-line-clamp: 3; // 行數
+                    -webkit-box-orient: vertical;
+                    white-space: normal;
+                }
+
+                .price {
+                    font-size: 16px;
+                    font-weight: 400;
+                    letter-spacing: 1px;
+                }
+            }
+        }
+
+        .choose {
+            margin: 10px 0 0 0;
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+
+            .quantity {
+                width: 120px;
+                display: flex;
+
+                .decrement {
+                    width: 30px;
+                    line-height: 30px;
+                    text-align: center;
+                    background: #B5B6B6;
+                    border-top-left-radius: 4px;
+                    border-bottom-left-radius: 4px;
+                    cursor: pointer;
+                    transition: 0.2s;
+
+                    &:hover {
+                        opacity: 0.8;
+                    }
+
+                    .minus {
+                        width: 10px;
+                        height: 10px;
+
+                        path {
+                            pointer-events: none;
+                        }
+                    }
+                }
+
+                .number {
+                    padding: 0;
+                    width: calc(100% - 30px * 2);
+                    height: 30px;
+                    background: #E4027E;
+                    text-align: center;
+                    border: none;
+                    border-radius: 0;
+                    outline: none;
+                    -moz-appearance: textfield;
+
+                    &::-webkit-outer-spin-button,
+                    &::-webkit-inner-spin-button {
+                        margin: 0;
+                        -webkit-appearance: none;
+                    }
+                }
+
+                .increment {
+                    width: 30px;
+                    line-height: 30px;
+                    text-align: center;
+                    background: #B5B6B6;
+                    border-top-right-radius: 4px;
+                    border-bottom-right-radius: 4px;
+                    cursor: pointer;
+                    transition: 0.2s;
+
+                    &:hover {
+                        opacity: 0.8;
+                    }
+
+                    .plus {
+                        width: 10px;
+                        height: 10px;
+
+                        path {
+                            pointer-events: none;
+                        }
+                    }
+                }
+            }
+
+            .total {
+                line-height: 30px;
+                color: #000;
+                font-size: 16px;
+                font-weight: 400;
+                letter-spacing: 1px;
+            }
+
+            .delete {
+                width: 30px;
+                height: 30px;
+                border-radius: 4px;
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                cursor: pointer;
+                transition: 0.2s;
+
+                &:hover {
+                    opacity: 0.8;
+
+                    .trash-can {
+                        fill: #E4027E;
+                    }
+                }
+
+                .trash-can {
+                    width: 24px;
+                    height: 24px;
+                    transition: 0.2s;
+
+                    path {
+                        pointer-events: none;
+                    }
+                }
+            }
+        }
+    }
+}
