+++ src/dataset/product.js
@@ -3,18 +3,21 @@ const PRODUCTS = [
         id: 1,
         image: require("../images/logo.png"),
         name: "元馨",
+        description: "一名手作創客，有一隻貓貓",
         price: 150,
     },
     {
         id: 2,
         image: require("../images/logo.png"),
         name: "元碩",
+        description: "一名前端工程師，喜歡彈鋼琴",
         price: 180,
     },
     {
         id: 3,
         image: require("../images/logo.png"),
         name: "阿蜜",
+        description: "一名超級阿嬤，很會交朋友",
         price: 300,
     }
 ]
